requirements.txt
asgiref==3.7.2
Django==4.2.11
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
